Update experiment header and card in bundles src
